Use float32 for loss and LoRA computations
Cast logits to float before calling F.cross_entropy and perform the LoRA path computation in float32 to avoid precision loss and overflow when tensors are in lower dtypes (e.g. fp16). The LoRA output is then converted back to the base output dtype to preserve model dtype. These changes improve numerical stability for mixed/half-precision training.

+++ mathllm/model/gpt2_arb.py
@@ -495,7 +495,7 @@ class TransformerWithARB(nn.Module):
         if valid_count.item() == 0:
             return flat_logits.new_zeros(())
         return F.cross_entropy(
-            flat_logits, flat_labels, ignore_index=-100, reduction="sum",
+            flat_logits.float(), flat_labels, ignore_index=-100, reduction="sum",
         ) / valid_count
 
     @torch.no_grad()
